@@ -1,31 +1,37 @@
 {
-  "name": "my-app",
-  "version": "0.1.0",
+  "name": "CivGui",
+  "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.8.2",
-    "@material-ui/icons": "^4.5.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/react": "^16.9.17",
-    "@types/react-dom": "^16.9.4",
-    "@types/react-router-dom": "^5.1.3",
+    "@emotion/react": "^11.5.0",
+    "@emotion/styled": "^11.3.0",
+    "@mui/icons-material": "^5.4.4",
+    "@mui/material": "^5.0.6",
+    "@mui/styles": "^5.0.2",
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^11.2.7",
+    "@testing-library/user-event": "^12.8.3",
+    "@types/axios": "^0.14.0",
+    "@types/jest": "^26.0.24",
+    "@types/node": "^12.20.36",
+    "@types/react": "^17.0.33",
+    "@types/react-dom": "^17.0.10",
+    "@types/react-redux": "^7.1.20",
+    "@types/react-router-dom": "^5.3.2",
     "@types/redux": "^3.6.0",
-    "axios": "^0.19.2",
-    "material-ui-confirm": "^2.0.4",
-    "react": "^16.12.0",
-    "react-confirm-alert": "^2.4.1",
-    "react-dom": "^16.12.0",
-    "react-draggable": "^4.4.3",
-    "react-localization": "^1.0.15",
-    "react-redux": "^7.1.3",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "3.3.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
-    "typescript": "^3.7.4",
-    "typescript-react-draggable": "^1.0.0"
+    "@types/redux-thunk": "^2.1.0",
+    "axios": "^0.21.4",
+    "material-ui-confirm": "^3.0.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-localization": "^1.0.17",
+    "react-redux": "^7.2.6",
+    "react-router-dom": "^5.3.0",
+    "react-scripts": "4.0.3",
+    "redux": "^4.1.2",
+    "ts-interface-checker": "^1.0.2",
+    "typescript": "^4.4.4",
+    "web-vitals": "^1.1.2"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -34,7 +40,10 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
@@ -49,6 +58,6 @@
     ]
   },
   "devDependencies": {
-    "@types/react-redux": "^7.1.7"
+    "ts-interface-builder": "^0.3.2"
   }
 }
