@@ -2,7 +2,7 @@ export { }
 
 import { FunctionComponent} from 'react'
 
-import { SvgIconProps } from '@material-ui/core/SvgIcon';
+import { SvgIconProps } from '@mui/material/SvgIcon';
 
 export interface TSvgComponent {
     props?: SvgIconProps,
